refactor(console): move uhyve code into `uhyve.rs`
