+++ src/console/mod.rs
@@ -1,3 +1,5 @@
+mod uhyve;
+
 use core::{fmt, mem};
 
 use embedded_io::{ErrorType, Read, ReadReady, Write};
@@ -9,12 +11,11 @@ use crate::arch::SerialDevice;
 use crate::drivers::console::VirtioUART;
 use crate::errno::Errno;
 use crate::executor::WakerRegistration;
-use crate::uhyve::serial_buf_hypercall;
 
 const SERIAL_BUFFER_SIZE: usize = 256;
 
 pub(crate) enum IoDevice {
-	Uhyve(UhyveSerial),
+	Uhyve(uhyve::UhyveSerial),
 	Uart(SerialDevice),
 	#[cfg(feature = "virtio-console")]
 	Virtio(VirtioUART),
@@ -62,42 +63,6 @@ impl Write for IoDevice {
 			crate::arch::kernel::vga::write_byte(byte);
 		}
 
-		Ok(buf.len())
-	}
-
-	fn flush(&mut self) -> Result<(), Self::Error> {
-		Ok(())
-	}
-}
-
-pub(crate) struct UhyveSerial;
-
-impl UhyveSerial {
-	pub const fn new() -> Self {
-		Self {}
-	}
-}
-
-impl ErrorType for UhyveSerial {
-	type Error = Errno;
-}
-
-impl Read for UhyveSerial {
-	fn read(&mut self, buf: &mut [u8]) -> Result<usize, Self::Error> {
-		let _ = buf;
-		Ok(0)
-	}
-}
-
-impl ReadReady for UhyveSerial {
-	fn read_ready(&mut self) -> Result<bool, Self::Error> {
-		Ok(false)
-	}
-}
-
-impl Write for UhyveSerial {
-	fn write(&mut self, buf: &[u8]) -> Result<usize, Self::Error> {
-		serial_buf_hypercall(buf);
 		Ok(buf.len())
 	}
 
@@ -185,7 +150,7 @@ pub(crate) static CONSOLE: Lazy<InterruptTicketMutex<Console>> = Lazy::new(|| {
 	crate::CoreLocal::install();
 
 	if crate::env::is_uhyve() {
-		return InterruptTicketMutex::new(Console::new(IoDevice::Uhyve(UhyveSerial::new())));
+		return InterruptTicketMutex::new(Console::new(IoDevice::Uhyve(uhyve::UhyveSerial::new())));
 	}
 
 	InterruptTicketMutex::new(Console::new(IoDevice::Uart(SerialDevice::new())))
